BE/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
idna==3.10
phonenumbers==9.0.4
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
pymongo==4.13.0